добавил обработку нулевого результата поиска

+++ StudentsTest/Controllers/StudentsController.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using StudentsTest.Entities;
 using StudentsTest.Services;
-
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+using System;
+using System.Threading.Tasks;
 
 namespace StudentsTest.Controllers
 {
@@ -15,11 +11,11 @@ namespace StudentsTest.Controllers
     public class StudentsController : ControllerBase
     {
         private readonly IStudentServices studentService;
-        public StudentsController(IStudentServices studentService) 
+        public StudentsController(IStudentServices studentService)
         {
             this.studentService = studentService;
         }
-        // GET: api/<StudentsController>
+
         [HttpGet("getAll")]
         public async Task<IActionResult> Get()
         {
@@ -33,29 +29,32 @@ namespace StudentsTest.Controllers
 
                 return BadRequest(ex.Message);
             }
-            
         }
 
-        // GET api/<StudentsController>/5
+
         [HttpGet("getById/{id}")]
         public async Task<IActionResult> GetById(int id)
         {
             try
             {
                 var result = await studentService.GetStudentById(id);
-                return Ok(result);
+                if (result != null) return Ok(result);
+                else return Ok("Данного студента не существует");
+
             }
             catch (Exception ex)
             {
 
                 return BadRequest(ex.Message);
-            }           
+            }
         }
         [HttpGet("deleteById/{id}")]
         public async Task<IActionResult> DeleteById(int id)
         {
             try
             {
+                var student = await studentService.GetStudentById(id);
+                if (student == null) return Ok("Данного студента не существует");
                 await studentService.DeleteStudent(id);
                 return Ok();
             }
@@ -64,15 +63,26 @@ namespace StudentsTest.Controllers
 
                 return BadRequest(ex.Message);
             }
-            
+
         }
 
         // POST api/<StudentsController>
-        [HttpPost]
-        public void Post([FromBody] string value)
+        [HttpPost("createOrUpdate")]
+        public async Task<IActionResult> CreateOrUpdateStudent([FromBody] StudentDTO student)
         {
+            try
+            {
+                //if ((await studentService.GetStudentById(student.Id))
+                var result = await studentService.CreateStudent(student);
+                return Ok(result);
 
+            }
+            catch (Exception ex)
+            {
+
+                return BadRequest(ex.Message);
+            }
         }
-       
+
     }
 }
